Add Scratch single-stage calculator prototype.
The Scratch version is simpler (works with a single stage and no frame-by-frame calculation), and was made as a prototype. It's functional (unlike the Python one currently), and has a graphical display of its output and graphing. This is the first version uploaded here and the first with the option of the Tsiolkovsky rocket equation (but subtract gravity acceleration times operating duration and multiply by payload mass) and better graphing (scaling and setting minimum, maximum and increment). It will be updated where I see improvement as well, it's not finished yet either.