Add option to graph with respect to fuel instead of payload, and equations (not yet applied) to find maximum effectual and optimal fuels for all outputs.
Sorry for inactivity, was finding necessary equations and some other internal changes.

Also added iterative Lambert W function (not sure which W value, I think W0, don't trust below -1/e (but you shouldn't input that anyway) but do trust below 1/e (the other one I found couldn't do that, this one can).

Lambert W is here because it's the same as the product log function WolframAlpha told me to use for the only answer it gave in the free version's computation time (which is the intersection at 0, the one I didn't want, but I resolved it now), redundant now but left in case someone wants to reparameterise it in the original equation (which always outputs 0).

Will make

I'm not sure how I'd add more buttons for these options when all 15 spaces are filled, possibly collapsibility or sub-menus (ie. a 'Thresholds' button for minimum TWR and maximum consumption or mass, possibly add pruning there (not enabled by default because a combination could be inferior enough to another to be pruned but)). But I like graphing having its own row in the main menu as an indicator when enabled, will keep it.

Will also add option for whether craft lifts off from the ground or the air (if gravity > 0, otherwise hide (also make minimum TWR hide as well (when inapplicable))), will use optimal fuel mass instead of actual (because all fuel above it will be burned while stationary).

This update also began differentiation between fuel tank and fuel mass, with converter function, will use in maximum profitable fuel calculation (not applicable in optimal, because the point at which additional fuel is detrimental is based on total craft mass, regardless of the proportion of which that's fuel, because they make no difference to acceleration when the additional fuel is being burned first).

Also, will eventually add mouse-only navigation (for mobile users, because they can use the buttons but can't toggle graphing mode or do the calculation).