Add electricity logistics, fix graphing bug
When electricity logistics are enabled, you can set your craft's built-in electricity generation and storage and the maximum number of solar panels allowed before it switches to adding RTGs (distribution between regular and arrays doesn't matter, their generation:mass ratio is the same).
It calculates the mass of the lightest possible such configuration for each amount generated from 0 to the net amount consumed, using batteries for the difference (but won't tell you to use batteries unless you have an excess of ≤~0.196t of fuel for panels or ~0.392t for RTGs).
Also, graphs are now a series of lines connecting points, not of vertical lines and delta-v is listed with each combination (more such variables will be added in the future).