Add the functions for calculating multistage craft delta-v, next will add interface
Many existing functions were generalised to avoid repetition, like the 'find engines in combination' function now has the combination string inputted, not the combinations list position, so it can also be used for finding combinations in possibilities (much of the programming for it is similar to the Python version, TurboWarp has sufficient performance to replace it and can be executed in browsers).
Also, this commit fixes electricity in regular mode, the mass of the lightest combination of necessary solar panels, batteries and RTGs is now correctly inputted to the Tsiolkovsky rocket equation, it wasn't in the previous one, not sure how far back this began but it works now.
Also, delta-v is calculated in the constants, not the non-constants, for now, because it uses the same loop, I'll organise it all soon, but it's currently functional, I think.