Allow stages to have different maximum engines in possibility generation (you don't need as many engines in a mostly-orbital stage as you do in a launch one).
To do this, add Index of Engine Quantity Increases in Combinations.
Also, thinking about using the engine superiority thresholds in gravity that I determined to speed it up by being less exhaustive, but I don't know how to account for all edge cases (where an otherwise-unreasonable number of a suboptimal engine type are better by merit of their granularity happening to line up with the optimal TWR). Also, it should stop calculating when the optimum one found stops increasing as you increase the number calculated (also calculable by finding whether the optimal's position in the combinations list is beneath the value of the final element of the new index of engine quantity increases).
Also, fun trivia: I initially considered using the same function for generating combinations out of engines across stages and possibilities out of combinations across multistage craft (and did in my first attempts, like how the engines in combination function and list are the same in both), but realised that counting in base 6 (or 7 with RCS thrusters) (but not numbers with digits greater than those succeeding them) is too different from counting in a different base for each digit.
Funner trivia: I like the 'item (item (target 1 + 1) of Stage Maximum Engines) of Index of Engine Quantity Increases in Combinations)' function, all nested list references are fun.