Fix exponentiation and prune function
I think this program needs to be rewritten from scratch if it's to work, too many things don't work and were made when I was less good at Python and didn't like multidimensional arrays as much. Perhaps it could begin from a port of the Newton-Raphson simultaneous solver, like the second Scratch one.